bench_util: add js_runtime benching utils
